tools/viz/requirements_viz.txt
blendify==2.1.0
trimesh
numpy
pillow
imageio
